Enable build server scenarios
Enable users to use SHFB and XsdDoc without actually installing it.
This is useful in build server scenarios where people want to include
small tools directly into the version control.

This has the nice side effect that we are now also able to bootstrap
the build i.e. we can use the version XsdDoc we just compiled for
building the documentation.