Change verbosity of log file to 'detailed'
The point of having a log file is that you have all the information
that you need if stuff goes wrong. It's currently set to 'normal'
which doesn't include an awful lot. We could also set it to
'diagnostic' but in my experience that makes it even harder to find
the stuff that's broken -- it suffers quite a bit of TMI.