Make `XsdPlugIn` compatible with SHFB v15.10.10
With SHFB v15.7.25 the `SandcastleProject` ctor used in this plugin has been removed. Here this ctor was used to create an intermediate project, for the purpose of merging conceptual content. According to the SHFB v15.7.25 release notes this approach is not needed any more.
Here I've used a custom `IContentFileProvider` and merged the new conceptual content using `BuildProcess.ConceptualContent`.